requirements.txt
fastapi
uvicorn
openai
python-dotenv
tinydb
langchain
langchain-core
pydantic
typing_extensions
langchain_openai
langchain_groq
logging
pymongo[svr]==4.13.0
requests
aiofiles
